Move AttributeValue to it's own file
