+++ scanamo/src/main/scala/com/gu/scanamo/query/ConditionExpression.scala
@@ -5,7 +5,7 @@ import com.gu.scanamo.DynamoFormat
 import com.gu.scanamo.error.{ConditionNotMet, ScanamoError}
 import com.gu.scanamo.ops.ScanamoOps
 import com.gu.scanamo.request.{RequestCondition, ScanamoDeleteRequest, ScanamoPutRequest, ScanamoUpdateRequest}
-import com.gu.scanamo.update.{AttributeName, UpdateExpression}
+import com.gu.scanamo.update.UpdateExpression
 import simulacrum.typeclass
 import cats.syntax.either._
