+++ scanamo/src/main/scala/com/gu/scanamo/query/DynamoKeyCondition.scala
@@ -2,7 +2,6 @@ package com.gu.scanamo.query
 
 import com.gu.scanamo.DynamoFormat
 import com.gu.scanamo.syntax.Bounds
-import com.gu.scanamo.update.AttributeName
 
 case class KeyEquals[V: DynamoFormat](key: Symbol, v: V) {
   def and[R: DynamoFormat](equalsKeyCondition: KeyEquals[R]) =
